local/requirements.txt
pandas
numpy
xgboost
matplotlib
scikit-learn
python-dotenv
joblib
statsmodels
requests
pytz
timezonefinder
fastapi
uvicorn